+++ Configuration_adv.hpp
@@ -226,10 +226,18 @@
 #endif
 
 #ifndef DEC_LIMIT_UP
-    #define DEC_LIMIT_UP 0.0f
+    #ifdef OAM
+        #define DEC_LIMIT_UP 100.0f
+    #else
+        #define DEC_LIMIT_UP 0.0f
+    #endif
 #endif
 #ifndef DEC_LIMIT_DOWN
-    #define DEC_LIMIT_DOWN 0.0f
+    #ifdef OAM
+        #define DEC_LIMIT_DOWN 100.0f
+    #else
+        #define DEC_LIMIT_DOWN 0.0f
+    #endif
 #endif
 
 ////////////////////////////
@@ -318,11 +326,10 @@
         #define AZ_STEPPER_ACCELERATION (100 * AZ_MICROSTEPPING)
     #endif
 
+    // the Circumference of the AZ rotation. 808mm dia.
     #ifndef AZ_CIRCUMFERENCE
-        // the Circumference of the AZ rotation. 808mm dia.
         #define AZ_CIRCUMFERENCE 2538.4f
     #endif
-
     #ifndef AZIMUTH_STEPS_PER_REV
         #define AZIMUTH_STEPS_PER_REV                                                                                                      \
             (AZ_CORRECTION_FACTOR * (AZ_CIRCUMFERENCE / (AZ_PULLEY_TEETH * GT2_BELT_PITCH)) * AZ_STEPPER_SPR                               \