+++ src/EPROMStore.hpp
@@ -47,12 +47,6 @@ class EEPROMStore
     static float getRollCalibrationAngle();
     static void storeRollCalibrationAngle(float rollCalibrationAngle);
 
-    static int32_t getRAParkingPos();
-    static void storeRAParkingPos(int32_t raParkingPos);
-
-    static int32_t getDECParkingPos();
-    static void storeDECParkingPos(int32_t decParkingPos);
-
     static float getDECLowerLimit();
     static void storeDECLowerLimit(float decLowerLimit);
 
@@ -64,6 +58,9 @@ class EEPROMStore
 
     static int32_t getDECHomingOffset();
     static void storeDECHomingOffset(int32_t decHomingOffset);
+
+    static int16_t getLastFlashedVersion();
+    static void storeLastFlashedVersion(int16_t lastVersion);
 
   private:
     /////////////////////////////////
@@ -88,11 +85,15 @@ class EEPROMStore
     // If Location 5 is 0xCF, then an extended 16-bit flag is stored in 21/22 and
     // indicates the additional fields that have been stored: 0000 0000 0000 0000
     //                                                        ^^^^ ^^^^ ^^^^ ^^^^
-    //                                                                       ||||
+    //                                                                  |||| ||||
+    //                   Last flashed version (56-57) ------------------+||| ||||
+    //                       DEC Homing Offet (52-55) -------------------+|| ||||
+    //      DEC Steps/deg, normalized to 256MS (48-51) -------------------+| ||||
+    //       RA Steps/deg, normalized to 256MS (44-47) --------------------+ ||||
     //                        RA Homing Offet (40-43) -----------------------+|||
     //                                UTC Offset (39) ------------------------+||
     //     DEC lower (31-34) and upper (35-38) limits -------------------------+|
-    //     RA (23-26) and DEC (27-30) Parking offsets --------------------------+
+    //     RA (23-26) and DEC (27-30) Parking offsets --------------------------+   ( ==== Obsolete V1.13.0 and beyond ==== )
     //
     /////////////////////////////////
 
@@ -125,6 +126,7 @@ class EEPROMStore
         RA_NORM_STEPS_MARKER_FLAG  = 0x0010,
         DEC_NORM_STEPS_MARKER_FLAG = 0x0020,
         DEC_HOMING_MARKER_FLAG     = 0x0040,
+        LAST_FLASHED_MARKER_FLAG   = 0x0080,
     };
 
     // These are the offsets to each item stored in the EEPROM
@@ -187,6 +189,8 @@ class EEPROMStore
         _DEC_HOMING_OFFSET_ADDR_1,
         _DEC_HOMING_OFFSET_ADDR_2,
         _DEC_HOMING_OFFSET_ADDR_3,  // Int32
+        LAST_FLASHED_VERSION = 56,
+        _LAST_FLASHED_VERSION_1,
         STORE_SIZE = 64
     };