+++ src/b_setup.hpp
@@ -12,6 +12,10 @@ PUSH_NO_WARNINGS
 POP_NO_WARNINGS
 #endif
 
+#ifndef NEW_STEPPER_LIB
+    #include "InterruptCallback.hpp"
+#endif
+
 #include "Utility.hpp"
 #include "EPROMStore.hpp"
 #include "a_inits.hpp"
@@ -65,10 +69,22 @@ void IRAM_ATTR stepperControlTask(void *payload)
     for (;;)
     {
         mountCopy->interruptLoop();
-        vTaskDelay(1);  // 1 ms 	// This will limit max stepping rate to 1 kHz
     }
 }
 
+#else
+    #ifndef NEW_STEPPER_LIB
+
+// This is the callback function for the timer interrupt on ATMega platforms.
+// It should do very minimal work, only calling Mount::interruptLoop() to step the stepper motors as needed.
+// It is called every 500 us (2 kHz rate)
+void stepperControlTimerCallback(void *payload)
+{
+    Mount *mountCopy = reinterpret_cast<Mount *>(payload);
+    if (mountCopy)
+        mountCopy->interruptLoop();
+}
+    #endif
 #endif
 
 /////////////////////////////////
@@ -305,6 +321,7 @@ void setup()
     LOG(DEBUG_ANY, "[STEPPERS]: Trk Microsteps  : %d", RA_TRACKING_MICROSTEPPING);
     LOG(DEBUG_ANY, "[STEPPERS]: Stepper SPR     : %d", RA_STEPPER_SPR);
     LOG(DEBUG_ANY, "[STEPPERS]: Transmission    : %f", RA_TRANSMISSION);
+    #ifdef NEW_STEPPER_LIB
     LOG(DEBUG_ANY, "[STEPPERS]: Driver Slew SPR : %l", config::Ra::DRIVER_SPR_SLEW);
     LOG(DEBUG_ANY, "[STEPPERS]: Driver Trk SPR  : %l", config::Ra::DRIVER_SPR_TRK);
     LOG(DEBUG_ANY, "[STEPPERS]: SPR Slew        : %f", config::Ra::SPR_SLEW);
@@ -313,10 +330,12 @@ void setup()
     LOG(DEBUG_ANY, "[STEPPERS]: Accel Slew      : %f", config::Ra::ACCEL_SLEW);
     LOG(DEBUG_ANY, "[STEPPERS]: Speed Trk       : %f", config::Ra::SPEED_TRK);
     LOG(DEBUG_ANY, "[STEPPERS]: Speed TrkComp   : %f", config::Ra::SPEED_COMPENSATION);
-
+    LOG(DEBUG_ANY, "[STEPPERS]: Configure RA stepper NEMA...");
     mount.configureRAStepper(RAmotorPin1, RAmotorPin2, config::Ra::SPEED_SLEW, config::Ra::ACCEL_SLEW);
-#else
-    #error New stepper type? Configure it here.
+    #else
+    LOG(DEBUG_ANY, "[STEPPERS]: Configure RA stepper NEMA...");
+    mount.configureRAStepper(RAmotorPin1, RAmotorPin2, RA_STEPPER_SPEED, RA_STEPPER_ACCELERATION);
+    #endif
 #endif
 
 #if (DEC_STEPPER_TYPE != STEPPER_TYPE_NONE)
@@ -324,6 +343,7 @@ void setup()
     LOG(DEBUG_ANY, "[STEPPERS]: Slew Microsteps : %d", DEC_SLEW_MICROSTEPPING);
     LOG(DEBUG_ANY, "[STEPPERS]: Stepper SPR     : %d", DEC_STEPPER_SPR);
     LOG(DEBUG_ANY, "[STEPPERS]: Transmission    : %f", DEC_TRANSMISSION);
+    #ifdef NEW_STEPPER_LIB
     LOG(DEBUG_ANY, "[STEPPERS]: Driver Slew SPR : %l", config::Dec::DRIVER_SPR_SLEW);
     LOG(DEBUG_ANY, "[STEPPERS]: Driver Trk SPR  : %l", config::Dec::DRIVER_SPR_TRK);
     LOG(DEBUG_ANY, "[STEPPERS]: SPR Slew        : %f", config::Dec::SPR_SLEW);
@@ -331,10 +351,12 @@ void setup()
     LOG(DEBUG_ANY, "[STEPPERS]: Speed Slew      : %f", config::Dec::SPEED_SLEW);
     LOG(DEBUG_ANY, "[STEPPERS]: Accel Slew      : %f", config::Dec::ACCEL_SLEW);
     LOG(DEBUG_ANY, "[STEPPERS]: Speed Trk       : %f", config::Dec::SPEED_TRK);
-
+    LOG(DEBUG_ANY, "[STEPPERS]: Configure DEC stepper NEMA...");
     mount.configureDECStepper(DECmotorPin1, DECmotorPin2, config::Dec::SPEED_SLEW, config::Dec::ACCEL_SLEW);
-#else
-    #error New stepper type? Configure it here.
+    #else
+    LOG(DEBUG_ANY, "[STEPPERS]: Configure DEC stepper NEMA...");
+    mount.configureDECStepper(DECmotorPin1, DECmotorPin2, DEC_STEPPER_SPEED, DEC_STEPPER_ACCELERATION);
+    #endif
 #endif
 
 #if RA_DRIVER_TYPE == DRIVER_TYPE_TMC2209_UART
@@ -422,6 +444,15 @@ void setup()
                             2,                   // Priority (2 is higher than 1)
                             &StepperTask,        // The location that receives the thread id
                             0);                  // The core to run this on
+
+#else
+    #ifndef NEW_STEPPER_LIB
+    // 2 kHz updates (higher frequency interferes with serial communications and complete messes up OATControl communications)
+    if (!InterruptCallback::setInterval(0.5f, stepperControlTimerCallback, &mount))
+    {
+        LOG(DEBUG_MOUNT, "[SYSTEM]: CANNOT setup interrupt timer!");
+    }
+    #endif
 #endif
 
 #if UART_CONNECTION_TEST_TX == 1
@@ -437,6 +468,9 @@ void setup()
     LOG(DEBUG_STEPPERS, "[STEPPERS]: Finished moving DEC axis using UART commands.");
     #endif
 #endif
+
+    LOG(DEBUG_ANY, "[SYSTEM]: Setting %s hemisphere...", inNorthernHemisphere ? "northern" : "southern");
+    mount.configureHemisphere(inNorthernHemisphere, true);
 
 #if TRACK_ON_BOOT == 1
     // Start the tracker.